Add files via upload
Post-Russian Invasion Survey of Smallholder Farmers in Ukraine
With a total agricultural area of close to 45 million ha of very fertile land,1 Ukraine has long been a 
major breadbasket and exporter of agricultural commodities. Beyond causing immediate hardship and 
triggering a large exodus of displaced people, Russia’s military invasion of Ukraine and the economic 
blockade of its Black Sea export routes have also led to sharp increases in grain prices (von CramonTaubadel 2022) and raised concern about global food security. To provide information to the 
government for developing policies and programs to support the agricultural sector in Ukraine after the 
war, the World Bank launched a nationwide survey of post-war farmers in cooperation with the Ministry 
of Agricultural Policy and Food (MAPF), with financial support from the EU, in areas controlled by 
Ukraine from October to December of 2022. 
The survey objective is to obtain information on changes in welfare, production, and productivity in the 
small and medium farm sector between 2021 and 2022 and to identify ways in which the farmers could 
be most effectively supported. The survey data is collected via phone by researchers at the Kyiv 
International Institute of Sociology (KIIS) under the monitoring of the research team in the World Bank